Dimitri createdeliverables

+++ ajax/deleteMember.php
@@ -2,6 +2,12 @@
 session_start();
 require_once($_SERVER['DOCUMENT_ROOT'] . '/includes/dbc.php');
 
+if(!WebUser::getUser()->isProf())
+{
+    exit("<p class='text-danger text-center'>You do not have enough privileges to delete a member from a group</p>");
+}
+
+
 $DeleteMember = new DeleteMember($_REQUEST['uid'], $_REQUEST['gid']);
 
 if($DeleteMember->delete())
